+++ ui/src/main/java/de/ipb_halle/lbac/exp/ExpRecordService.java
@@ -35,11 +35,11 @@ import de.ipb_halle.lbac.admission.User;
 import de.ipb_halle.lbac.exp.assay.AssayService;
 import de.ipb_halle.lbac.exp.text.TextService;
 import de.ipb_halle.lbac.items.Item;
-import de.ipb_halle.lbac.items.UnknownItemFactory;
+import de.ipb_halle.lbac.items.InaccessibleItemFactory;
 import de.ipb_halle.lbac.items.service.ItemService;
 import de.ipb_halle.lbac.material.Material;
 import de.ipb_halle.lbac.material.common.service.MaterialService;
-import de.ipb_halle.lbac.material.unknown.UnknownMaterial;
+import de.ipb_halle.lbac.material.inaccessible.InaccessibleMaterial;
 
 import java.io.Serializable;
 import java.util.ArrayList;
@@ -221,14 +221,14 @@ public class ExpRecordService implements Serializable {
             if (e.getItemId() != null) {
                   item = this.itemService.loadItemById(e.getItemId());
                 if (!aclistService.isPermitted(ACPermission.permREAD, item, user)) {
-                    item = UnknownItemFactory.getInstance(user, GlobalAdmissionContext.getPublicReadACL());
+                    item = InaccessibleItemFactory.getInstance(user, GlobalAdmissionContext.getPublicReadACL());
                 } 
                 material = item.getMaterial();
             } else {
                 if (e.getMaterialId() != null) {
                     material = this.materialService.loadMaterialById(e.getMaterialId());
                     if (!aclistService.isPermitted(ACPermission.permREAD, material, user)) {
-                        material = UnknownMaterial.createNewInstance(GlobalAdmissionContext.getPublicReadACL());
+                        material = InaccessibleMaterial.createNewInstance(GlobalAdmissionContext.getPublicReadACL());
 
                     }
                 }
