+++ ui/src/main/java/de/ipb_halle/lbac/material/MaterialType.java
@@ -60,7 +60,7 @@ public enum MaterialType implements Serializable {
             6, MaterialDetailType.COMMON_INFORMATION),
     TAXONOMY(
             7, MaterialDetailType.COMMON_INFORMATION),
-    UNKNOWN(8);
+    INACCESSIBLE(8);
 
     private final List<MaterialDetailType> types;
     private static final Map<String, MaterialType>  string2Enum = Stream.of(values()).collect(toMap(Object::toString, e -> e));
